// src/lib/ApiTypes.d.ts
export interface components {
	schemas: {
		LoginUser: {
			email: string;
			/** Format: password */
			password: string;
		};
		NewUser: {
			username: string;
			email: string;
			/** Format: password */
			password: string;
		};
		User: {
			email: string;
			token: string;
			username: string;
			bio: string;
			image: string;
		};
		UpdateUser: {
			email?: string;
			password?: string;
			username?: string;
			bio?: string;
			image?: string;
		};
		Profile: {
			username: string;
			bio: string;
			image: string;
			following: boolean;
		};
		Article: {
			slug: string;
			title: string;
			description: string;
			body: string;
			tagList: string[];
			/** Format: date-time */
			createdAt: string;
			/** Format: date-time */
			updatedAt: string;
			favorited: boolean;
			favoritesCount: number;
			author: components["schemas"]["Profile"];
		};
		NewArticle: {
			title: string;
			description: string;
			body: string;
			tagList?: string[];
		};
		UpdateArticle: {
			title?: string;
			description?: string;
			body?: string;
		};
		Comment: {
			id: number;
			/** Format: date-time */
			createdAt: string;
			/** Format: date-time */
			updatedAt: string;
			body: string;
			author: components["schemas"]["Profile"];
		};
		NewComment: {
			body: string;
		};
		GenericErrorModel: {
			errors: {
				body: string[];
			};
		};
	};
	responses: {
		/** @description Tags */
		TagsResponse: {
			content: {
				"application/json": {
					tags: string[];
				};
			};
		};
		/** @description Single comment */
		SingleCommentResponse: {
			content: {
				"application/json": {
					comment: components["schemas"]["Comment"];
				};
			};
		};
		/** @description Multiple comments */
		MultipleCommentsResponse: {
			content: {
				"application/json": {
					comments: components["schemas"]["Comment"][];
				};
			};
		};
		/** @description Single article */
		SingleArticleResponse: {
			content: {
				"application/json": {
					article: components["schemas"]["Article"];
				};
			};
		};
		/** @description Multiple articles */
		MultipleArticlesResponse: {
			content: {
				"application/json": {
					articles: components["schemas"]["Article"][];
					articlesCount: number;
				};
			};
		};
		/** @description Profile */
		ProfileResponse: {
			content: {
				"application/json": {
					profile: components["schemas"]["Profile"];
				};
			};
		};
		/** @description User */
		UserResponse: {
			content: {
				"application/json": {
					user: components["schemas"]["User"];
				};
			};
		};
		/** @description No content */
		EmptyOkResponse: {
			content: {};
		};
		/** @description Unauthorized */
		Unauthorized: {
			content: {};
		};
		/** @description Unexpected error */
		GenericError: {
			content: {
				"application/json": components["schemas"]["GenericErrorModel"];
			};
		};
	};
	parameters: {
		/** @description The number of items to skip before starting to collect the result set. */
		offsetParam: number;
		/** @description The numbers of items to return. */
		limitParam: number;
	};
	requestBodies: {
		/** @description Credentials to use */
		LoginUserRequest: {
			content: {
				"application/json": {
					user: components["schemas"]["LoginUser"];
				};
			};
		};
		/** @description Details of the new user to register */
		NewUserRequest: {
			content: {
				"application/json": {
					user: components["schemas"]["NewUser"];
				};
			};
		};
		/** @description User details to update. At least **one** field is required. */
		UpdateUserRequest: {
			content: {
				"application/json": {
					user: components["schemas"]["UpdateUser"];
				};
			};
		};
		/** @description Article to create */
		NewArticleRequest: {
			content: {
				"application/json": {
					article: components["schemas"]["NewArticle"];
				};
			};
		};
		/** @description Article to update */
		UpdateArticleRequest: {
			content: {
				"application/json": {
					article: components["schemas"]["UpdateArticle"];
				};
			};
		};
		/** @description Comment you want to create */
		NewCommentRequest: {
			content: {
				"application/json": {
					comment: components["schemas"]["NewComment"];
				};
			};
		};
	};
	headers: never;
	pathItems: never;
}